Move Jira client to new jbi.jira.client module
